Modifications for delft upgrade without the delft reactor design

+++ tally/meshTally.cxx
@@ -2,8 +2,8 @@
   CombLayer : MNCPX Input builder
  
  * File:   tally/meshTally.cxx
-*
- * Copyright (c) 2004-2013 by Stuart Ansell
+ *
+ * Copyright (c) 2004-2014 by Stuart Ansell
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -167,10 +167,11 @@ meshTally::setIndex(const int* IDX)
 {
   ELog::RegMethod RegA("meshTally","setIndex");
 
-  for(int i=0;i<3;i++)
+  for(size_t i=0;i<3;i++)
     {
       if (IDX[i]<1)
 	throw ColErr::IndexError<int>(IDX[i],1,RegA.getBase());
+
       Pts[i]=IDX[i];
     }
   return;
@@ -312,7 +313,7 @@ meshTally::write(std::ostream& OX) const
 	  cx<<"mshmf"<<IDnum<<" "<<mshmf;
 	  StrFunc::writeMCNPX(cx.str(),OX);
 	}
-      for(int i=0;i<3;i++)
+      for(size_t i=0;i<3;i++)
 	OX<<"cor"<<abc[i]<<IDnum<<" "<<minCoord[i]<<" "
 	  <<Pts[i]<<"i "<<maxCoord[i]<<std::endl;
       OX<<"endmd"<<std::endl;
